+++ src/identity_mapper/responses.py
@@ -1,11 +1,13 @@
 from identity_mapper.capability_protocol import (
     AuthenticateResponse,
+    MapIdentityResponse,
     ResolveIdentityResponse,
     VerifyCredentialResponse,
 )
 
 __all__ = [
     "AuthenticateResponse",
+    "MapIdentityResponse",
     "ResolveIdentityResponse",
     "VerifyCredentialResponse",
 ]